Time: 153 ms (69.50%) | Memory: 107.9 MB (47.16%) - LeetSync

+++ 3351-maximize-happiness-of-selected-children/maximize-happiness-of-selected-children.cpp
@@ -6,8 +6,7 @@ public:
         sort(happiness.begin(),happiness.end());
         for(int i=0;i<k;i++){
             int val=happiness[n-i-1]-i;
-            //if(val < 0) val=0;
-            sum+=max(0,val);
+            sum+=max(0LL,(long long)val);
         }
         return sum;
     }
